world/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==3.2
django-cors-headers==4.3.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.4.1
djangorestframework==3.13.1
json-parser==1.2.0
jsons==1.6.3
pillow==10.3.0
pycparser==2.22
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
typish==1.9.3
